requirements.txt
numpy
openai
faiss-cpu
python-dotenv